move configureStore into helpers
