+++ webpack-assets.json
@@ -1,6 +1,6 @@
 {
   "javascript": {
-    "main": "http://localhost:3001/dist/main-505b5a3b4bdc1c010a8a.js"
+    "main": "http://localhost:3001/dist/main-45a4f85bbed044d9aaaa.js"
   },
   "styles": {},
   "assets": {
